+++ homebrew/maccrab.rb
@@ -1,5 +1,5 @@
 cask "maccrab" do
-  version "1.6.15"
+  version "1.6.16"
   sha256 "fac1bd386fe896d0bcfa1eabb91eb9eebdf00d4f1f58be7cfa4946a71f4d8e06"
 
   url "https://github.com/peterhanily/maccrab/releases/download/v#{version}/MacCrab-v#{version}.dmg"